+++ packages/create-mechanic/function-templates/svg-video/index.js
@@ -1,4 +1,4 @@
-export const handler = ({ inputs, mechanic }) => {
+export const handler = ({ inputs, frame, done, drawLoop }) => {
   const { width, height, text, color1, color2, radiusPercentage, turns } =
     inputs;
 
@@ -6,7 +6,7 @@ export const handler = ({ inputs, mechanic }) => {
   const radius = ((height / 2) * radiusPercentage) / 100;
 
   let angle = 0;
-  const drawFrame = () => {
+  drawLoop(() => {
     const svg = `<svg width="${width}" height="${height}">
       <rect fill="#F4F4F4" width="${width}" height="${height}" />
       <g transform="translate(${center[0]}, ${center[1]})">
@@ -36,65 +36,62 @@ export const handler = ({ inputs, mechanic }) => {
     </svg>`;
 
     if (angle < turns * 360) {
-      mechanic.frame(svg);
+      frame(svg);
       angle += 360 / 100;
-      window.requestAnimationFrame(drawFrame);
     } else {
-      mechanic.done(svg);
+      done(svg);
     }
-  };
-
-  drawFrame();
+  });
 };
 
 export const inputs = {
   width: {
     type: "number",
-    default: 400,
+    default: 400
   },
   height: {
     type: "number",
-    default: 300,
+    default: 300
   },
   text: {
     type: "text",
-    default: "mechanic",
+    default: "mechanic"
   },
   color1: {
     type: "color",
     model: "hex",
-    default: "#E94225",
+    default: "#E94225"
   },
   color2: {
     type: "color",
     model: "hex",
-    default: "#002EBB",
+    default: "#002EBB"
   },
   radiusPercentage: {
     type: "number",
     default: 40,
     min: 0,
     max: 100,
-    slider: true,
+    slider: true
   },
   turns: {
     type: "number",
-    default: 3,
-  },
+    default: 3
+  }
 };
 
 export const presets = {
   medium: {
     width: 800,
-    height: 600,
+    height: 600
   },
   large: {
     width: 1600,
-    height: 1200,
-  },
+    height: 1200
+  }
 };
 
 export const settings = {
   engine: require("@mechanic-design/engine-svg"),
-  animated: true,
+  animated: true
 };